fix circular import bug
