+++ rdagent/scenarios/data_science/proposal/exp_gen/refine.py
@@ -6,8 +6,11 @@ from rdagent.core.proposal import ExpGen
 from rdagent.log import rdagent_logger as logger
 from rdagent.oai.llm_utils import APIBackend
 from rdagent.scenarios.data_science.experiment.experiment import DSExperiment
-from rdagent.scenarios.data_science.proposal.exp_gen import CodingSketch, get_component
 from rdagent.scenarios.data_science.proposal.exp_gen.base import DSHypothesis, DSTrace
+from rdagent.scenarios.data_science.proposal.exp_gen.utils import (
+    CodingSketch,
+    get_component,
+)
 from rdagent.utils.agent.tpl import T
